pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lea"]

[project]
name = "lea-prover"
version = "0.1.0"
description = "A minimal Lean 4 theorem proving agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.70.0",
]

[project.scripts]
lea = "lea.cli:main"
